+++ lib/dart_holodex_api.dart
@@ -9,4 +9,5 @@ export 'package:intl/locale.dart' show Locale;
 export 'src/enums.dart';
 export 'src/exception.dart';
 export 'src/holodex_client.dart';
+export 'src/holodex_endpoint.dart';
 export 'src/models.dart';